Evasive impl by calculating point distance

+++ src/main/java/map/Maze.java
@@ -65,7 +65,7 @@ public class Maze {
         return Maze.specialPastilleCounter;
     }
 
-    public boolean isValidLocation(int line, int column) throws Exception {
+    public boolean isValidLocation(double line, double column) throws Exception {
         //TODO correct condition
         if (line == 0 && column == 0) {
             throw new Exception(String.format("Invalid Location X: %s Y: %s", column, line));
